refactor(auth): extract and utilize `useAuthToken` hook for token management
- Moved `useAuthToken` hook into a dedicated file (`useAuthToken.ts`).
